Fix location group activity stream links
The activity stream implementation isn't easily extendable for custom
group types, so we are monkey patching the code which generates the
group links in the activity stream so the correct group controller is
used.

The group_link helper only seems to be used in activity streams so this
should be pretty safe, and perhaps could be generalised if we want to
fix the activity stream on events (it seems we have disabled it
currently).